refactor: moves filterapi under internal (#1179)
**Description**

Previously, we put the configuration API for the AI Gateawy external
processor as a public package because initially the ExtProc was designed
to be able to work independent of the control plane ~v0.1. That was true
only until before v0.2 where we fully started leveraging Envoy Gateway's
extension server as well as the two phase filter architecture to be able
to do more robust fallbacks, etc. As a result, it is nearly impossible
to make it work outside our control plane, so there's no reason to keep
the filterapi package in public.

---------
