+++ internal/controller/gateway_test.go
@@ -26,8 +26,8 @@ import (
 	"sigs.k8s.io/yaml"
 
 	aigv1a1 "github.com/envoyproxy/ai-gateway/api/v1alpha1"
-	"github.com/envoyproxy/ai-gateway/filterapi"
 	"github.com/envoyproxy/ai-gateway/internal/controller/rotators"
+	"github.com/envoyproxy/ai-gateway/internal/filterapi"
 )
 
 func TestGatewayController_Reconcile(t *testing.T) {
