+++ internal/extproc/backendauth/auth.go
@@ -11,7 +11,7 @@ import (
 
 	extprocv3 "github.com/envoyproxy/go-control-plane/envoy/service/ext_proc/v3"
 
-	"github.com/envoyproxy/ai-gateway/filterapi"
+	"github.com/envoyproxy/ai-gateway/internal/filterapi"
 )
 
 // Handler is the interface that deals with the backend auth for a specific backend.
